Removed all calls to exit() in common/src.
I realize now that having exit() there only causes problems.  The two
functions that called exit(), set_signal() and read_dtime() have been
changed as follows:

set_signal(): returns -1 on error, or 0 if okay.  Don't bother checking the
return code in code, since if it does return an error, there's a problem
with the code, not a general error (class 1---see

	http://boston.conman.org/2009/12/01.2

for more information), but you might want to check the system log if you
add calls to set_signal().

read_dtime(): pass in a default value that's returned when the given
argument is invalid.