Changed location of an assert, removed some code no longer needed
I made it so that a NULL base pointer can be passed in to bisearch(), and
as long as the nelem is 0, no harm, no foul.  This required the assert
checking for a non-NULL base to be moved into the main while() loop.

Also, I removed the NULL check in edomain_search() as it's no longer
required for proper operation.