Bug fix---fix TCP control messages timing bug
The code had a timing issue between the server and client---sometimes (and
it varied from system to system) the TCP socket for various reports wasn't
ready by the time the client attempted to connect, as the "go ahead" message
was sent to the client too early.  Moved the "go ahead" reply to after the
full TCP configuration is done.